frontend: helpers: Extract cluster settings stuff to clusterSettings
