+++ frontend/src/components/common/NamespacesAutocomplete.tsx
@@ -9,7 +9,8 @@ import React, { useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { useHistory, useLocation } from 'react-router-dom';
-import helpers, { addQuery } from '../../helpers';
+import { addQuery } from '../../helpers';
+import { loadClusterSettings } from '../../helpers/clusterSettings';
 import { useCluster, useClustersConf } from '../../lib/k8s';
 import Namespace from '../../lib/k8s/namespace';
 import { setNamespaceFilter } from '../../redux/filterSlice';
@@ -129,7 +130,7 @@ export function NamespacesAutocomplete() {
   const [namespaceNames, setNamespaceNames] = React.useState<string[]>([]);
 
   React.useEffect(() => {
-    const settings = helpers.loadClusterSettings(cluster || '');
+    const settings = loadClusterSettings(cluster || '');
     const allowedNamespaces = settings?.allowedNamespaces || [];
     if (allowedNamespaces.length > 0) {
       setNamespaceNames(allowedNamespaces);
